Crate wasmtime_versioned_export_macros

This crate defines macros to easily define and use functions with a versioned suffix, to facilitate using multiple versions of the same crate that generate assembly.

Macros

versioned_stringify_ident

versioned_suffix

Attribute Macros

versioned_export

versioned_link